Move plugin discovery into plugins service
Move plugin post-connect discovery execution out of gestaltd/internal/discovery and into services/plugins/discovery.

Interface diff:

- import "github.com/valon-technologies/gestalt/server/internal/discovery"

- discovery.Run(ctx, cfg, client)

+ import "github.com/valon-technologies/gestalt/server/services/plugins/discovery"

+ discovery.Run(ctx, cfg, client)

Example:

candidates, err := discovery.Run(ctx, provider.DiscoveryConfig(), client)
